Use get_object_or_404() instead of  ObjectDoesNotExist
Why do we use a helper function get_object_or_404() instead of
automatically catching the ObjectDoesNotExist exceptions at a higher
level, or having the model API raise Http404 instead of
ObjectDoesNotExist?

Because that would couple the model layer to the view layer. One of the
foremost design goals of Django is to maintain loose coupling. Some
controlled coupling is introduced in the django.shortcuts module.